raspberry-kinect-cultivo-growth-estimation/requirements.txt
spade
freenect
opencv-python
numpy
pandas
paho-mqtt
RPi.GPIO
